requirements.txt
pytorch-lightning==2.0.2
PyYAML==6.0
tensorboard>=2.2.0
torch>=1.6.1
torchsummary==1.5.1
torchvision>=0.10.1
tqdm
# unittest==0.0
flask
Pillow
scipy